+++ examples/unix-seqpacket-client.rs
@@ -30,15 +30,7 @@ async fn do_main(options: &Options) -> Result<(), String> {
 		.await
 		.map_err(|e| format!("failed to send request: {}", e))?;
 
-	loop {
-		let update = request
-			.recv_update()
-			.await
-			.map_err(|e| format!("failed to read message: {}", e))?;
-		let update = match update {
-			Some(x) => x,
-			None => break,
-		};
+	while let Some(update) = request.recv_update().await {
 		let message = std::str::from_utf8(&update.body.data).map_err(|_| "invalid UTF-8 in update")?;
 		eprintln!("Received update: {}", message);
 	}