+++ src/request.rs
@@ -26,6 +26,7 @@ pub struct ReceivedRequest<Body> {
 	service_id: i32,
 	incoming_rx: mpsc::UnboundedReceiver<Message<Body>>,
 	command_tx: mpsc::UnboundedSender<Command<Body>>,
+	peek_buffer: Option<Message<Body>>,
 }
 
 /// An incoming request or stream message.
@@ -66,15 +67,15 @@ impl<Body> SentRequest<Body> {
 
 	/// Receive the next update message of the request from the remote peer.
 	///
-	/// This function returns `Ok(None)` if the final response is received instead of an update message.
+	/// This function returns `None` if the final response is received instead of an update message.
 	/// If that happens, the response message can be read using [`Self::recv_response`].
-	pub async fn recv_update(&mut self) -> Result<Option<Message<Body>>, error::RecvMessageError> {
+	pub async fn recv_update(&mut self) -> Option<Message<Body>> {
 		let message = self.recv_message().await?;
 		if message.header.message_type.is_responder_update() {
-			Ok(Some(message))
+			Some(message)
 		} else {
 			self.peek_buffer = Some(message);
-			Ok(None)
+			None
 		}
 	}
 
@@ -84,7 +85,7 @@ impl<Body> SentRequest<Body> {
 	/// If that happens, the update message can be read using [`Self::recv_update`].
 	/// To ensure that there are no update messages left, keep calling [`Self::recv_update`] untill it returns `Ok(None)`.
 	pub async fn recv_response(&mut self) -> Result<Message<Body>, error::RecvMessageError> {
-		let message = self.recv_message().await?;
+		let message = self.recv_message().await.ok_or_else(error::connection_aborted)?;
 		let kind = message.header.message_type;
 		if kind.is_response() {
 			Ok(message)
@@ -100,11 +101,11 @@ impl<Body> SentRequest<Body> {
 	/// Receive the next message of the request from the remote peer.
 	///
 	/// This could be an update message or a response message.
-	async fn recv_message(&mut self) -> Result<Message<Body>, error::RecvMessageError> {
+	async fn recv_message(&mut self) -> Option<Message<Body>> {
 		if let Some(message) = self.peek_buffer.take() {
-			Ok(message)
+			Some(message)
 		} else {
-			Ok(self.incoming_rx.recv().await.ok_or_else(error::connection_aborted)?)
+			self.incoming_rx.recv().await
 		}
 	}
 
@@ -148,6 +149,7 @@ impl<Body> ReceivedRequest<Body> {
 			service_id,
 			incoming_rx,
 			command_tx,
+			peek_buffer: None,
 		}
 	}
 
@@ -162,8 +164,12 @@ impl<Body> ReceivedRequest<Body> {
 	}
 
 	/// Receive the next update message of the request from the remote peer.
-	pub async fn recv_update(&mut self) -> Result<Message<Body>, error::ReadMessageError> {
-		Ok(self.incoming_rx.recv().await.ok_or_else(error::connection_aborted)?)
+	pub async fn recv_update(&mut self) -> Option<Message<Body>> {
+		if let Some(message) = self.peek_buffer.take() {
+			Some(message)
+		} else {
+			self.incoming_rx.recv().await
+		}
 	}
 
 	/// Send an update for the request to the remote peer.
@@ -194,6 +200,19 @@ impl<Body> ReceivedRequest<Body> {
 			.send(SendRawMessage { message, result_tx }.into())
 			.map_err(|_| error::connection_aborted())?;
 		result_rx.await.map_err(|_| error::connection_aborted())?
+	}
+
+	/// Put a message back in the peek buffer.
+	///
+	/// Do not use this function.
+	/// It is not convered by the version number API stability guarantee.
+	///
+	/// # Panics
+	/// This function panics if there already is a message in the peek buffer.
+	#[doc(hidden)]
+	pub fn _unpeek_message(&mut self, message: Message<Body>) {
+		assert!(self.peek_buffer.is_none());
+		self.peek_buffer = Some(message);
 	}
 }